requirements.txt
matplotlib>=3.5.0
pydantic>=2.0.0
uvicorn>=0.20.0
fastapi>=0.100.0
starlette>=0.27.0
sse-starlette>=1.6.0
mcp>=0.9.0
PyJWT>=2.8.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
